Refactoring so all handlers are in individual classes and messages have a type
Needed to change to a ‘type’ rather than message[‘message’] and make it
upper case to match expectations.
